Organize socket tests by file name.
