+++ test/tests/app/socket/middleware/test-auth-middleware.js
@@ -1,18 +1,16 @@
-const knex = require('../../../../app/lib/db');
+const knex = require('../../../../../app/lib/db');
 const chai = require('chai');
 const chaiHttp = require('chai-http');
-const server = require('../../../../app');
-const API_TEAM_LOGIN_URL = require('../../../data/constants').API_TEAM_LOGIN_URL;
-const authenticationMiddleware = require('../../../../app/socket/middleware/authentication');
-const roomController = require('../../../../app/socket/controllers/room');
+const server = require('../../../../../app');
+const API_TEAM_LOGIN_URL = require('../../../../data/constants').API_TEAM_LOGIN_URL;
+const authenticationMiddleware = require('../../../../../app/socket/middleware/authentication');
 const randomString = require('randomstring');
-const timeout = require('../../../data/constants').TIMEOUT;
 
 chai.use(chaiHttp);
 const should = chai.should();
 let teamToken = null;
 
-describe('Team Socket Functions', () => {
+describe('Socket AuthenticationMiddleware Tests', () => {
   before((done) => {
     // Authenticate a Team to perform protected queries.
     function authenticateTeam() {
@@ -54,7 +52,7 @@ describe('Team Socket Functions', () => {
     return mockSocket;
   }
 
-  it('Should succeed validate Team Token', (done) => {
+  it('Should succeed validateTeamToken', (done) => {
     const socket = getMockSocket();
     socket.emit = (eventName, eventData) => {
       eventName.should.equal('info');
@@ -65,7 +63,7 @@ describe('Team Socket Functions', () => {
     authenticationMiddleware.validateTeamToken(socket, teamToken, () => {});
   });
 
-  it('Should fail validate Team Token', (done) => {
+  it('Should fail validateTeamToken', (done) => {
     const invalidTeamToken = 'THIS_IS_FAKE';
     const socket = getMockSocket();
     socket.emit = (eventName, eventData) => {
@@ -76,24 +74,4 @@ describe('Team Socket Functions', () => {
     };
     authenticationMiddleware.validateTeamToken(socket, invalidTeamToken, () => {});
   });
-
-  it('Should join a Socket to a Room', (done) => {
-    const socket = getMockSocket();
-    socket.to = () => socket;
-    socket.join = (room) => {
-      should.exist(room);
-    };
-    socket.emit = (eventName, eventData) => {
-      eventName.should.equal('info');
-      eventData.event.should.equal('join room');
-      eventData.didSucceed.should.equal(true);
-    };
-    roomController.joinRoom(socket, 'game_room');
-
-    // Delay for emitter checks.
-    setTimeout(() => {
-      done();
-    }, timeout);
-  });
 });
-
